refactor: 폴더명 store를 storage로 변경
